server done

+++ interface/src/settings.rs
@@ -46,7 +46,7 @@ pub struct TgCollector {
     pub database_directory: String,
     pub log_verbosity_level: i32,
     pub encryption_key: String,
-    pub api_id: i64,
+    pub api_id: i32,
     pub api_hash: String,
     pub phone: String,
     pub max_download_queue_size: usize,
